Create main.js
This code receives a reference to our square div, then sets a variable called angle to 0. We then use the setInterval() method to increase the value of angle by 2 (we also use the % operator so that it resets to 0 at 360), then set the transform CSS3 property to rotate that number of degrees. The second argument is 1000/60 , which equates to a frame speed of 60 frames per second.